requirements.txt
python-dotenv==0.21.0
google-cloud-dialogflow==2.14.1
googletrans==3.1.0a0
pydub==0.25.1
SpeechRecognition==3.8.1
python-telegram-bot==13.8.1
openai==0.25.0
dalle2==1.0.11
